Avoid code duplication: use bytestring_is_valid_utf8 on any arch without vectorised instructions
