+++ bytestring.cabal
@@ -122,13 +122,10 @@ library
   c-sources:        cbits/fpstring.c
                     cbits/itoa.c
   
-  if (arch(x86_64) || arch(i386))
-      c-sources:        cbits/x86/is-valid-utf8.c
+  if (arch(aarch64))
+    c-sources:        cbits/aarch64/is-valid-utf8.c
   else
-    if (arch(aarch64))
-      c-sources:        cbits/aarch64/is-valid-utf8.c
-    else
-      c-sources:        cbits/is-valid-utf8.c
+    c-sources:        cbits/is-valid-utf8.c
   
   cc-options:        -std=c11
